+++ client/index.html
@@ -3,13 +3,30 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Top Movies</title>
+    <title>Topins Toop</title>
     <link rel="stylesheet" href="./public/styles/scss/pico.slim.scss">
+    <link rel="apple-touch-icon" sizes="57x57" href="./Public/PWA/apple-icon-57x57.png">
+    <link rel="apple-touch-icon" sizes="60x60" href="./Public/PWA/apple-icon-60x60.png">
+    <link rel="apple-touch-icon" sizes="72x72" href="./Public/PWA/apple-icon-72x72.png">
+    <link rel="apple-touch-icon" sizes="76x76" href="./Public/PWA/apple-icon-76x76.png">
+    <link rel="apple-touch-icon" sizes="114x114" href="./Public/PWA/apple-icon-114x114.png">
+    <link rel="apple-touch-icon" sizes="120x120" href="./Public/PWA/apple-icon-120x120.png">
+    <link rel="apple-touch-icon" sizes="144x144" href="./Public/PWA/apple-icon-144x144.png">
+    <link rel="apple-touch-icon" sizes="152x152" href="./Public/PWA/apple-icon-152x152.png">
+    <link rel="apple-touch-icon" sizes="180x180" href="./Public/PWA/apple-icon-180x180.png">
+    <link rel="icon" type="image/png" sizes="192x192" href="./Public/PWA/android-icon-192x192.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="./Public/PWA/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="96x96" href="./Public/PWA/favicon-96x96.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="./Public/PWA/favicon-16x16.png">
+    <link rel="manifest" href="./Public/PWA/manifest.json">
+    <meta name="msapplication-TileColor" content="#ffffff">
+    <meta name="msapplication-TileImage" content="./Public/PWA/ms-icon-144x144.png">
+    <meta name="theme-color" content="#ffffff">
     <!-- <script defer type="module" src="/scripts/titles.js"></script> -->
     <link rel="stylesheet" href="./public/styles/sass/styles.scss">
   </head>
-  <body id="app" class="container" style="background-color:#11191f;">
-
+  <body class="container">
+    <div id="app"></div>
   </body>
-  <script type="module" src="./main.jsx"></script>
+  <script type="module" src="./src/main.jsx"></script>
 </html>
