+++ client/public/components/dom-states/loading.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+const Loading = () => {
+  return (
+    <section>
+      <article>
+        <h1>Loading...</h1>
+      </article>
+    </section>
+  )
+}
+export default Loading;
