--- client/public/components/index.js
@@ -1,6 +1,0 @@
-import { Header } from "./header";
-import { Nav } from "./nav";
-import { Footer } from "./footer";
-import { Card } from "./card";
-import { ErrorPage } from './error-page';
-export { Header, Nav, Footer, Card, ErrorPage };
