--- client/public/components/nav.jsx
@@ -1,42 +1,0 @@
-import React from 'react';
-import { NavLink } from 'react-router-dom';
-const Nav = ()  => {
-  return (
-    <nav className="">
-      <menu>
-        <li>
-          <strong>
-            <NavLink 
-              to={`.`}
-              className={({
-                isActive
-              }) =>  isActive ? '
-              ' : ''}
-              style={{
-                color: 'aqua'
-              }}
-            >
-              Top Movies
-            </NavLink>
-          </strong>
-        </li>
-      </menu>
-      <menu>
-        <li>
-          <NavLink 
-            className={({isActive}) =>  {
-
-            }}
-            style={{
-
-            }}  
-          >
-            Home
-          </NavLink>
-        </li>
-      </menu>
-    </nav>
-  )
-};
-
-export default Nav;
