+++ client/public/components/semantics/footer.jsx
@@ -19,5 +19,4 @@ const Footer = () => {
       </footer>
   )
 };
-
 export default Footer;
