+++ client/public/components/semantics/header.jsx
@@ -5,5 +5,4 @@ const Header = () => {
     <h1>Top Movies</h1>
   </header>
 )};
-
 export default Header;
