+++ client/public/styles/sass/styles.scss
@@ -140,6 +140,14 @@ body {
   background: transparent;
 }
 
+.cards {
+  display: flex;
+  flex-wrap: wrap;
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
+
 #footer-container {
   position: fixed;
   width: 100%;
