+++ client/src/main.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App.jsx';
+
 ReactDOM.createRoot(document.getElementById('app')).render(
   <React.StrictMode>
       <>
