+++ client/src/pages/TopMovies.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { Card } from '../../public/components/card'
+const TopMovies = () => {
+  return (
+    <section className={`cards`}>
+      <Card />
+    </section>
+  )
+}
+export default TopMovies;
